Implement a very first version of our twitter client used to extract smiles
