+++ src/main/scala/com/github/pedrovgs/haveaniceday/smiles/apiclient/TwitterClient.scala
@@ -1,14 +1,37 @@
 package com.github.pedrovgs.haveaniceday.smiles.apiclient
 
+import com.github.pedrovgs.haveaniceday.extensions.date._
 import com.danielasfregola.twitter4s.TwitterRestClient
-import com.github.pedrovgs.haveaniceday.smiles.model.SmilesExtractionResult
+import com.danielasfregola.twitter4s.entities.{RatedData, Tweet}
+import com.github.pedrovgs.haveaniceday.smiles.model.{Smile, SmilesExtractionResult, Source, UnknownError}
 import org.joda.time.DateTime
+
 import scala.concurrent.Future
+import scala.concurrent.ExecutionContext.Implicits.global
 
 class TwitterClient {
 
-  def smilesFrom(extractionDate: DateTime): Future[SmilesExtractionResult] = ???
-
   private val restClient = TwitterRestClient()
 
+  def smilesFrom(accounts: Seq[String], extractionDate: DateTime): Future[SmilesExtractionResult] = {
+    val eventualSmiles = accounts.map { account =>
+      restClient
+        .userTimelineForUserId(account.toLong, exclude_replies = true, include_rts = false, count = Int.MaxValue)
+        .flatMap { timeline =>
+          Future.successful(toSmiles(extractionDate, timeline))
+        }
+    }
+    Future.sequence(eventualSmiles).map(_.flatten).map(Right(_)).recover {
+      case e => Left(UnknownError(e.getMessage))
+    }
+  }
+
+  private def toSmiles(extractionDate: DateTime, timeline: RatedData[Seq[Tweet]]): Seq[Smile] = {
+    val validTweets = timeline.data.filter(_.created_at.compareTo(extractionDate.toDate) < 0)
+    validTweets.map { tweet =>
+      val url   = "TODO"
+      val photo = tweet.extended_entities.flatMap(_.media.headOption.map(_.display_url))
+      Smile(tweet.id, tweet.created_at, photo, Some(tweet.text), Source.Twitter, url, sent = false, None, None)
+    }
+  }
 }
